requirements.txt
pandas==1.0.1
python-dotenv==0.9.1
pymongo==3.7.2
flask
dnspython